I dont know if I want to do this shit anymore
Dude I have gotten out of the honey moon phase of this. Like I have no clue what I'm doing and we have 4 weeks till comp and we have a frame and an elevator that has been waiting to be mounted for A WEEK like shit is it that hard. We are so cooked for mag why even go at this point I honestly am so tired of this stuff the people are great but this is driving me crazy. but who gives a crap no one reads this and that pathplaner  dependency is so dumb and my mentor expects the robot to have April tag tracking and shit but I barley know how to make an elevator. Like most of these problems are skill issues and I have no idea why I wanted to be a programmer in the first place I do not like doing this I much prefer to talk to people and not just sit in front of a computer all day looking at words. ig this is the end of my rant I kinda just want to quit and go sit someplace quite and just exist for a bit.