after bayou push
well bayou did not go well. The last time we had gone un picked at a regional was in 2014 so the whole team is kinda feeling like shit rn. I personally think we did better than 2024 but I also know we got picked at both comps. I know that the biggest issue was coral getting stuck at the bottom of our elevator, that is a fatal design flaw that was killing us. But this design flaw had a simple solution, tech-no-tigers did it after their first comp and it worked it is to put fabric at the bottom of the lower rung of the elevator and to the frame. I brought this up to our lead coach and he said no, why you may ask I have no idea because he never gave me a reason. He instead told us the great advance of to "Just do not miss". I again personally think that most of the fault for us losing is prob on myself and on Mr Young. Me because I was not able to get auto line up working and on Mr Young for not allowing us to do a simple fix and for over all being a massive downer. I understand that we were losing but I do not think that as a coach he should be putting us down that much, he had called me the worst programmer in wildcat robotics history many times and to be fair he is kinda right but in no world do you say that to someone right into their face while they are the only person who can do the most important job on the team. I know this is been a kinda long rant so im going to end it here.