HUGE changes for HUGE results
Well it works now and I'm prob going to push this to main for real!

HUGE thanks to 4336 / p-ry because of their pose estimator  and also to my dog Fortez for helping me learn how to code in the first place. Next challenge is to get mt2 working (no big sweat right... right?)